Api Tags factored out into separate file
