+++ src/main/scala/code/api/builder/APIMethods_APIBuilder.scala
@@ -2,6 +2,7 @@ package code.api.builder
 import java.util.UUID
 import code.api.builder.JsonFactory_APIBuilder._
 import code.api.util.APIUtil._
+import code.api.util.ApiTag._
 import code.api.util.ApiVersion
 import code.api.util.ErrorMessages._
 import net.liftweb.common.Full
