+++ src/main/scala/code/api/util/APIUtil.scala
@@ -46,6 +46,7 @@ import code.api.OAuthHandshake._
 import code.api.UKOpenBanking.v2_0_0.OBP_UKOpenBanking_200
 import code.api.berlin.group.v1.OBP_BERLIN_GROUP_1
 import code.api.oauth1a.Arithmetics
+import code.api.util.ApiTag.ResourceDocTag
 import code.api.util.Glossary.GlossaryItem
 import code.api.v1_2.ErrorMessage
 import code.api.{DirectLogin, util, _}
@@ -1003,54 +1004,6 @@ object APIUtil extends MdcLoggable {
    */
 
 
-  // Used to tag Resource Docs
-  case class ResourceDocTag(tag: String)
-
-  // Use the *singular* case. for both the variable name and string.
-  // e.g. "This call is Payment related"
-  val apiTagTransactionRequest = ResourceDocTag("Transaction-Request")
-  val apiTagApi = ResourceDocTag("API")
-  val apiTagBank = ResourceDocTag("Bank")
-  val apiTagAccount = ResourceDocTag("Account")
-  val apiTagAccountPublic = ResourceDocTag("Account-Public")
-  val apiTagAccountFirehose = ResourceDocTag("Account-Firehose")
-  val apiTagFirehoseData = ResourceDocTag("FirehoseData")
-  val apiTagPublicData = ResourceDocTag("PublicData")
-  val apiTagPrivateData = ResourceDocTag("PrivateData")
-  val apiTagTransaction = ResourceDocTag("Transaction")
-  val apiTagTransactionFirehose = ResourceDocTag("Transaction-Firehose")
-  val apiTagCounterpartyMetaData = ResourceDocTag("Counterparty-Metadata")
-  val apiTagTransactionMetaData = ResourceDocTag("Transaction-Metadata")
-  val apiTagView = ResourceDocTag("Account-View")
-  val apiTagEntitlement = ResourceDocTag("Entitlement")
-  val apiTagRole = ResourceDocTag("API-Role")
-  val apiTagOwnerRequired = ResourceDocTag("OwnerViewRequired")
-  val apiTagCounterparty = ResourceDocTag("Counterparty")
-  val apiTagKyc = ResourceDocTag("KYC")
-  val apiTagCustomer = ResourceDocTag("Customer")
-  val apiTagOnboarding = ResourceDocTag("Onboarding")
-  val apiTagUser = ResourceDocTag("User")
-  val apiTagMeeting = ResourceDocTag("Customer-Meeting")
-  val apiTagExperimental = ResourceDocTag("Experimental")
-  val apiTagPerson = ResourceDocTag("Person")
-  val apiTagCard = ResourceDocTag("Card")
-  val apiTagSandbox = ResourceDocTag("API-Sandbox")
-  val apiTagBranch = ResourceDocTag("Bank-Branch")
-  val apiTagATM = ResourceDocTag("Bank-ATM")
-  val apiTagProduct = ResourceDocTag("Bank-Product")
-  val apiTagOpenData = ResourceDocTag("Open-Data")
-  val apiTagConsumer = ResourceDocTag("API-Consumer")
-  val apiTagSearchWarehouse = ResourceDocTag("Data-Warehouse")
-  val apiTagFx = ResourceDocTag("Bank-FX")
-  val apiTagMessage = ResourceDocTag("Customer-Message")
-  val apiTagMetric = ResourceDocTag("API-Metric")
-  val apiTagDocumentation = ResourceDocTag("API-Documentation")
-  val apiTagBerlinGroup = ResourceDocTag("Berlin-Group")
-  val apiTagUKOpenBanking = ResourceDocTag("UKOpenBanking")
-  val apiTagApiBuilder = ResourceDocTag("API_Builder")
-  val apiTagAggregateMetrics = ResourceDocTag("Aggregate-Metrics")
-  val apiTagNewStyle = ResourceDocTag("New-Style")
-  val apiTagWebHook = ResourceDocTag("Web-Hook")
 
   case class Catalogs(core: Boolean = false, psd2: Boolean = false, obwg: Boolean = false)
