+++ src/main/scala/code/api/v2_1_0/APIMethods210.scala
@@ -8,6 +8,7 @@ import code.api.util
 import code.api.util.ErrorMessages.TransactionDisabled
 import code.api.util.NewStyle.HttpCode
 import code.api.util.{APIUtil, ApiRole}
+import code.api.util.ApiTag._
 import code.api.v1_2_1.AmountOfMoneyJsonV121
 import code.api.v1_3_0.{JSONFactory1_3_0, _}
 import code.api.v1_4_0.JSONFactory1_4_0
