+++ src/main/scala/code/api/v2_2_0/APIMethods220.scala
@@ -6,6 +6,7 @@ import java.util.{Date, Locale, UUID}
 import code.actorsystem.ObpActorConfig
 import code.api.ResourceDocs1_4_0.SwaggerDefinitionsJSON._
 import code.api.util.APIUtil._
+import code.api.util.ApiTag._
 import code.api.util.ApiRole._
 import code.api.util.ErrorMessages.{BankAccountNotFound, _}
 import code.api.util.{APIUtil, ApiRole, ApiVersion, ErrorMessages}
